fix: use amount of rules when generating randoms

+++ src/components/rule_display.rs
@@ -2,8 +2,8 @@ use crate::RULES;
 use leptos::prelude::*;
 
 #[component]
-pub fn RuleDisplay(#[prop(into)] rule: Signal<u32>) -> impl IntoView {
-    let rule = Signal::derive(move || RULES.get(rule.get() as usize).unwrap().clone());
+pub fn RuleDisplay(#[prop(into)] rule: Signal<usize>) -> impl IntoView {
+    let rule = Signal::derive(move || RULES.get(rule.get()).unwrap().clone());
 
     view! {
         <div class="w-full max-w-md p-6">
